Refactor to have message display handled by background.js
The content script is reduced to forwarding the keyEvents to background.js and to the displayMessage function used to display various messages.

Note that this means that from now on, logmeout.js is significantly different from the bookmarklet since it has message passing with the background page and doesn't manage the message display.